WebAssembly-Design.md fixes (#84)
* Add missing paren.

* Slightly improve wording.

* Use natural notation for magic.

As if it read byte-by-byte.

* Bytecode is one word.

* Replace SHOULD with MUST.

This capitalized keyword looks like from commonly accepted https://www.ietf.org/rfc/rfc2119.txt for indication requirement levels.

The following statements are requirements, not recomendation. Thus the keyword should be not SHOULD but MUST.

* deterministic stach height validation → metering

Also add an empty vertical space between the text and the list.