remove the public address and private key
https://etherscan.io/address/0x00a329c0648769A73afAc7F9381E08FB43dBEA72

this address is listened by hack and if the developer forget to remove the account from metamask after development. once the mainnet money is
transfered to this address, the hack will transfer the money to somewhere else. 

There are so many eth the hacker have been success stolen, I lost 1 eth this morning:
https://etherscan.io/tx/0x7041400e301abc855832d20a5350783078df298e6fe14fb092ef1d9304f53ed9

faucet is a better solution than this!